clear up more compiler warnings
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1003906 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/analysis/common/src/java/org/apache/lucene/analysis/compound/hyphenation/TernaryTree.java
@@ -258,7 +258,7 @@ public class TernaryTree implements Cloneable, Serializable {
       }
     }
     if (a[start + i] != 0) {
-      return (int) -a[start + i];
+      return -a[start + i];
     }
     return 0;
